Force sig_init_lrecruits[j] <- 0.3
This parameter has been problematic, but it is a nuisance parameter. We won't ever use it for inference, and it will not be used in forward simulations. It is used only in state initialization and when trying to estimate it the posterior looks too much like the prior, suggesting it is not estimable. I think assuming this parameter is known is fine, we could even do a sensitivity analysis to see if it changes anything later on.