Include winter tagging data in jags_data
Placed in the Lphi_obs_Pa_Ma object. This was previously used only for fall migrant survival from the trap to LGR. The "spring-mig" dimension was previously all NAs, but now the winter tagging data are placed there.

This sort of assumes that negligible mortality occurs between trap tagging and winter tagging