Fix bug in calculation of Yr1 ocean surv qresids
The old approach calculated the deviation from the all year-mean based on the total time series SD. This includes both stochastic and non-stochastic parts.

The new approach is calculated based on the deviation from the expected value for the given year, after accounting for the prior year's deviate. In this way it quantifies only the deviation from that year's posterior predictive distribution, not that of any given year.